Split cli.py into focused modules for readability and maintenance
Extract three new modules from the monolithic cli.py (~1000 lines):

- magnet_type.py: MagnetType enum, DEFAULT_* constants, validate_magnet_config
- measures.py: build_patterns, create_measure_dict, build_power/heat/temperature/
  mechanical measures, configure_magnet_target/postvalues, loadMdata
- export.py: exportResults

cli.py now only contains main() (~120 lines).
commissioning.py updated to import loadMdata/exportResults from the new modules
directly instead of from cli.
